requirements/base.txt
dj-rest-auth==2.1.7
Django[argon2]==3.2.5
django-environ==0.4.5
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
